requirements.txt
torch>=2.0
diffusers>=0.28.0
transformers>=4.38.0
accelerate>=0.28.0
wandb>=0.16.0
datasets>=2.18.0
ccxt>=4.2.0
pandas>=2.0.0
numpy>=1.24.0
mplfinance>=0.12.10
Pillow>=10.0.0
pandas-ta>=0.3.14
matplotlib>=3.8.0
tqdm>=4.66.0
safetensors>=0.4.0
